backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
faiss-cpu==1.11.0.post1
whoosh==2.7.4
sentence-transformers==5.1.0
numpy==2.2.6
pydantic==2.11.7
